feat(Ban): Don't send the message twice when ban/unbanning
feat(Ban): Don't send the ban message twice when ran in #support-bans

feat(Unban): Don't send the message twice when ran in #support-bans

See also: #25
