+++ src/lib/mod.rs
@@ -1,4 +1,4 @@
 pub mod cooldown;
+pub mod hypixel;
 pub mod interact_opts;
 pub mod print_embeds;
-pub mod hypixel;
